Fix non-destructive pipethrough of source when no sourcemap exists
When no source map exists in the input file, sourceMolder.comment is set to
null. Therefor, when queue is called on line 91, the needle for str.replace is
`null` which gets string cast into `"null"`. That cases the first instance of
the `null` token in the source code to get replaced with where ever was passed
into the callback.

This commit includes a test to replicate the behavior and a fix. The fix works
by checking for the existence of the sourcemap comment before running the
replace functionality.